Revert "Towers and minions have been implemented ( though not tested yet ). Some useless classes have been removed."
This reverts commit 42cac447493e40b7d8334f8133201abfcd5a47ed.

+++ Medieval_POST-APOCALYPSE_logica/src/mpa/core/logic/AbstractObject.java
@@ -68,10 +68,10 @@ public abstract class AbstractObject implements InterfaceAbstractObject
 		this.width = width;
 	}
 
-	public float getCollisionRay()
+	public double getCollisionRay()
 	{
 		float xMin = x - width / 2;
 		float yMin = y - height / 2;
-		return ( float ) Math.sqrt( Math.pow( ( xMin - x ), 2 ) + Math.pow( ( yMin - y ), 2 ) );
+		return Math.sqrt( Math.pow( ( xMin - x ), 2 ) + Math.pow( ( yMin - y ), 2 ) );
 	}
 }
